MVP completed

+++ src/main/java/Passenger.java
@@ -1,10 +1,10 @@
 public class Passenger {
 
     private String name;
-    private int contactNumber;
+    private long contactNumber;
     private int id;
 
-    public Passenger(String name, int contactNumber, int id){
+    public Passenger(String name, long contactNumber, int id){
         this.name = name;
         this.contactNumber = contactNumber;
         this.id = id;
